Move JsonTestHelper to testutil
