+++ document/pom.xml
@@ -48,11 +48,6 @@
             <scope>test</scope>
         </dependency>
         <dependency>
-            <groupId>uk.co.datumedge</groupId>
-            <artifactId>hamcrest-json</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
             <groupId>com.yahoo.vespa</groupId>
             <artifactId>config</artifactId>
             <version>${project.version}</version>
@@ -82,6 +77,12 @@
         <artifactId>guice</artifactId>
         <classifier>no_aop</classifier>
         <scope>provided</scope>
+      </dependency>
+      <dependency>
+          <groupId>com.yahoo.vespa</groupId>
+          <artifactId>testutil</artifactId>
+          <version>${project.version}</version>
+          <scope>test</scope>
       </dependency>
     </dependencies>
     <build>
