+++ document/src/main/java/com/yahoo/document/json/JsonWriter.java
@@ -112,6 +112,7 @@ public class JsonWriter implements DocumentWriter {
                 Map.Entry<Field, FieldValue> entry = i.next();
                 entry.getValue().serialize(entry.getKey(), this);
             }
+
             generator.writeEndObject();
             generator.writeEndObject();
             generator.flush();
