+++ testutil/src/main/java/com/yahoo/test/json/JsonTestHelper.java
@@ -1,15 +1,15 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.config.subscription.util;
+package com.yahoo.test.json;
 
 import com.google.common.base.Joiner;
+import org.hamcrest.MatcherAssert;
 
-import static org.hamcrest.MatcherAssert.assertThat;
 import static uk.co.datumedge.hamcrest.json.SameJSONAs.sameJSONAs;
 
 /**
  * @author Vegard Sjonfjell
  */
-public class JsonHelper {
+public class JsonTestHelper {
     /**
      * Convenience method to input JSON without escaping double quotes and newlines
      * Each parameter represents a line of JSON encoded data
@@ -23,6 +23,6 @@ public class JsonHelper {
      * Structurally compare two JSON encoded strings
      */
     public static void assertJsonEquals(String inputJson, String expectedJson) {
-        assertThat(inputJson, sameJSONAs(expectedJson));
+        MatcherAssert.assertThat(inputJson, sameJSONAs(expectedJson));
     }
 }
